+++ afkak/test/test_group.py
@@ -14,19 +14,48 @@
 # limitations under the License.
 from __future__ import absolute_import
 
+import logging
+import pprint
+
 from mock import Mock, patch
 from twisted.internet import defer, task
 from twisted.python.failure import Failure
 from twisted.trial import unittest
 
 from afkak.common import (
-    ConsumerCoordinatorNotAvailableError, IllegalGenerationError,
-    InconsistentGroupProtocolError, JoinGroupRequestProtocol,
-    JoinGroupResponseMember, NotCoordinatorForConsumerError,
-    RebalanceInProgressError, RequestTimedOutError, RestartError, RestopError,
-    SyncGroupMemberAssignment, UnknownError, UnknownMemberIdError,
+    ConsumerCoordinatorNotAvailableError, HeartbeatResponse,
+    IllegalGenerationError, InconsistentGroupProtocolError,
+    JoinGroupRequestProtocol, JoinGroupResponse, JoinGroupResponseMember,
+    NotCoordinatorForConsumerError, RebalanceInProgress, RequestTimedOutError,
+    RestartError, RestopError, SyncGroupMemberAssignment, SyncGroupResponse,
+    UnknownError, UnknownMemberIdError,
 )
 from afkak.group import ConsumerGroup, ConsumerProtocol, Coordinator
+
+log = logging.getLogger(__name__)
+
+
+def assert_delayed_calls(n, client):
+    """
+    :param int n: Expected number of delayed calls.
+    :param client:
+        An object like `afkak.client.KafkaClient` which has a *reactor*
+        attribute.
+    """
+    calls = client.reactor.getDelayedCalls()
+    calls_repr = pprint.pformat([vars(c) for c in calls])
+    log.debug('assert_delayed_calls(%d) found calls %s', n, calls_repr)
+    if len(calls) == n:
+        return
+    raise AssertionError((
+        "Expected {expected} delayed calls at time {now},"
+        " but found {actual}:\n{calls}"
+    ).format(
+        expected=n,
+        actual=len(calls),
+        now=client.reactor.seconds(),
+        calls=calls_repr,
+    ))
 
 
 class Base(unittest.TestCase):
@@ -40,13 +69,20 @@ class Base(unittest.TestCase):
         return Coordinator(client, "group_id", ["topic1"], protocol_cls=Mock())
 
     def join_response(self, member_id="m1", leader_id="m1"):
-        return defer.succeed(Mock(
-            error_code=0, member_id=member_id, generation_id="g1",
-            leader_id=leader_id, members=[]))
+        return defer.succeed(JoinGroupResponse(
+            error=0,
+            generation_id="g1",
+            group_protocol="consumer",
+            member_id=member_id,
+            leader_id=leader_id,
+            members=[],
+        ))
 
     def sync_response(self):
-        return defer.succeed(Mock(
-            error_code=0, member_assignment=[]))
+        return defer.succeed(SyncGroupResponse(
+            error=0,
+            member_assignment=[],
+        ))
 
 
 class TestCoordinator(Base):
@@ -62,7 +98,7 @@ class TestCoordinator(Base):
 
     def test_send_join_group_request_failure(self):
         client = self.mock_client([
-            defer.fail(RebalanceInProgressError()),
+            defer.fail(RebalanceInProgress()),
         ])
         coord = self.make_coordinator(client)
         de = coord.send_join_group_request()
@@ -82,7 +118,7 @@ class TestCoordinator(Base):
 
     def test_send_sync_group_request_failure(self):
         client = self.mock_client([
-            defer.fail(RebalanceInProgressError()),
+            defer.fail(RebalanceInProgress()),
         ])
         coord = self.make_coordinator(client)
         de = coord.send_sync_group_request([])
@@ -121,7 +157,8 @@ class TestCoordinator(Base):
             Successfully join and sync as follower
         """
         client = self.mock_client([
-            self.join_response(leader_id="m2"), self.sync_response()
+            self.join_response(leader_id="m2"),
+            self.sync_response(),
         ])
         coord = self.make_coordinator(client)
         de = coord.join_and_sync()
@@ -132,14 +169,15 @@ class TestCoordinator(Base):
             Get an error when joining and retry
         """
         client = self.mock_client([
-            defer.fail(RebalanceInProgressError()),
-            self.join_response(), self.sync_response()
+            defer.fail(RebalanceInProgress()),
+            self.join_response(),
+            self.sync_response(),
         ])
         coord = self.make_coordinator(client)
         de = coord.join_and_sync()
         self.successResultOf(de)
         self.assertEqual(coord._rejoin_needed, True)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 1)
+        assert_delayed_calls(1, client)
         de = coord.join_and_sync()
         self.successResultOf(de)
         self.assertEqual(coord._rejoin_needed, False)
@@ -150,14 +188,15 @@ class TestCoordinator(Base):
         """
         client = self.mock_client([
             self.join_response(),
-            defer.fail(RebalanceInProgressError()),
-            self.join_response(), self.sync_response()
+            defer.fail(RebalanceInProgress()),
+            self.join_response(),
+            self.sync_response(),
         ])
         coord = self.make_coordinator(client)
         de = coord.join_and_sync()
         self.successResultOf(de)
         self.assertEqual(coord._rejoin_needed, True)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 1)
+        assert_delayed_calls(1, client)
         de = coord.join_and_sync()
         self.successResultOf(de)
         self.assertEqual(coord._rejoin_needed, False)
@@ -207,28 +246,36 @@ class TestCoordinator(Base):
             run successful heartbeats, get a resync message, and schedule resync
         """
         client = self.mock_client([
-            self.join_response(), self.sync_response(),
-            defer.succeed(Mock(error_code=0)),
-            defer.fail(RebalanceInProgressError()),
-            self.join_response(), self.sync_response(),
+            self.join_response(),
+            self.sync_response(),
+            defer.succeed(HeartbeatResponse(error=0)),
+            defer.fail(RebalanceInProgress()),
+            self.join_response(),
+            self.sync_response(),
         ])
         coord = self.make_coordinator(client)
-        de = coord.join_and_sync()
-        self.successResultOf(de)
+        self.successResultOf(coord.join_and_sync())
         self.assertEqual(coord._rejoin_needed, False)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 1)
-        coord._heartbeat()
+        assert_delayed_calls(1, client)
+
+        # Run the heartbeat, which succeeds.
+        client.reactor.advance(5.0)
         self.assertEqual(coord._rejoin_needed, False)
         self.assertFalse(coord._heartbeat_request_d)
-        coord._heartbeat()
+
+        # Run a heartbeat, which fails with RebalanceInProgress
+        client.reactor.advance(5.0)
+        self.assertEqual(coord._rejoin_needed, True)
         self.assertFalse(coord._heartbeat_request_d)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 2)
+        # There is a delayed call to join_and_sync():
+        assert_delayed_calls(1, client)
         self.assertEqual(coord._rejoin_needed, True)
-        coord._heartbeat()
-        self.assertEqual(coord._rejoin_needed, True)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 2)
-        de = coord.join_and_sync()
-        self.successResultOf(de)
+
+        client.reactor.advance(0.1)  # Call join_and_sync()
+        self.assertEqual(coord._rejoin_needed, False)
+        assert_delayed_calls(1, client)  # Heartbeat scheduled.
+
+        self.assertIs(None, coord.join_and_sync())
         self.assertEqual(coord._rejoin_needed, False)
 
     def test_heartbeat_coordinator_lost(self):
@@ -236,17 +283,19 @@ class TestCoordinator(Base):
             run a heartbeat that indicates the coordinator has changed
         """
         client = self.mock_client([
-            self.join_response(), self.sync_response(),
+            self.join_response(),
+            self.sync_response(),
             defer.fail(NotCoordinatorForConsumerError()),
         ])
         coord = self.make_coordinator(client)
-        de = coord.join_and_sync()
-        self.successResultOf(de)
+        self.successResultOf(coord.join_and_sync())
+
         self.assertEqual(coord._rejoin_needed, False)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 1)
-        coord._heartbeat()
+
+        assert_delayed_calls(1, client)
+        client.reactor.advance(6.0)
         self.assertFalse(coord._heartbeat_request_d)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 2)
+        assert_delayed_calls(1, client)
         self.assertEqual(coord._rejoin_needed, True)
         client.reset_consumer_group_metadata.assert_called_with(coord.group_id)
 
@@ -255,17 +304,19 @@ class TestCoordinator(Base):
             get a heartbeat timer tick while a request is already in progress
         """
         client = self.mock_client([
-            self.join_response(), self.sync_response(),
-            defer.succeed(Mock(error_code=0)),
+            self.join_response(),
+            self.sync_response(),
+            defer.succeed(HeartbeatResponse(error=0)),
         ])
         coord = self.make_coordinator(client)
-        de = coord.join_and_sync()
-        self.successResultOf(de)
+        self.successResultOf(coord.join_and_sync())
         self.assertEqual(coord._rejoin_needed, False)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 1)
+
+        assert_delayed_calls(1, client)  # Heartbeat scheduled.
+
         coord._heartbeat_request_d = defer.Deferred()
         coord.send_heartbeat_request = Mock()
-        coord._heartbeat()
+        client.reactor.advance(5.0)
         coord.send_heartbeat_request.assert_not_called()
         coord._heartbeat_request_d = None
         coord._stopping = True
@@ -340,7 +391,7 @@ class TestCoordinator(Base):
         client._get_brokerclient.return_value = None
         de = coord.get_coordinator_broker()
         self.successResultOf(de)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 1)
+        self.assertEqual(1, len(client.reactor.getDelayedCalls()))
 
         client._get_coordinator_for_group.return_value = defer.succeed(Mock())
         client._get_brokerclient.return_value = coord
@@ -362,23 +413,24 @@ class TestCoordinator(Base):
         client._get_brokerclient.return_value = None
         de = coord.get_coordinator_broker()
         self.successResultOf(de)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 1)
-        client._get_coordinator_for_group.return_value = defer.fail(
-            UnknownError())
+        assert_delayed_calls(1, client)  # Heartbeat scheduled.
+        client._get_coordinator_for_group.return_value = defer.fail(UnknownError())
         de = coord.get_coordinator_broker()
         self.successResultOf(de)
 
     def test_get_coordinator_fatal(self):
         """
-        fail to retrieve a coordinator in a non-retriable way
+        A non-retriable failure when retrieving the coordinator broker leaves
+        the coordinator in a quiescent state.
         """
         client = self.mock_client([])
+        client._get_coordinator_for_group.return_value = defer.fail(AttributeError())
         coord = self.make_coordinator(client)
-        client._get_coordinator_for_group.return_value = defer.fail(
-            AttributeError())
-        de = coord.get_coordinator_broker()
-        self.failureResultOf(de)
-        self.assertEqual(client._get_clock.return_value.callLater.call_count, 0)
+
+        self.failureResultOf(coord.get_coordinator_broker()).check(AttributeError)
+
+        # No heartbeat scheduled.
+        self.assertEqual([], client.reactor.getDelayedCalls())
 
     def test_rejoin_after_error(self):
         """
@@ -391,21 +443,21 @@ class TestCoordinator(Base):
         def check(rejoin_needed, exc):
             coord._rejoin_needed = False
             coord._rejoin_wait_dc = None
-            callLater = client._get_clock.return_value.callLater
-            callLater.reset_mock()
+            for call in client.reactor.getDelayedCalls():
+                call.cancel()
             client.reset_consumer_group_metadata.reset_mock()
             coord.on_group_leave.reset_mock()
 
             coord.rejoin_after_error(Failure(exc))
             if rejoin_needed:
                 self.assertEqual(coord._rejoin_needed, True)
-                self.assertEqual(callLater.call_count, 1)
+                assert_delayed_calls(1, client)
             else:
                 self.assertEqual(coord._rejoin_needed, False)
-                self.assertEqual(callLater.call_count, 0)
+                assert_delayed_calls(0, client)
                 self.assertEqual(coord._rejoin_wait_dc, None)
 
-        check(True, RebalanceInProgressError())
+        check(True, RebalanceInProgress())
         check(True, ConsumerCoordinatorNotAvailableError())
         client.reset_consumer_group_metadata.assert_any_call(coord.group_id)
         check(True, IllegalGenerationError())